Rename `helpers` folder to `config`
